requirements.txt
absl-py==0.12.0
astunparse==1.6.3
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
chess==1.4.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-SocketIO==4.3.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
imgviz==1.2.6
itsdangerous==1.1.0
Jinja2==2.11.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme==4.5.7
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
more-itertools==8.7.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.1
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-chess==1.999
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.2
Shapely==1.7.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2021.4.8
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
